Split definitions
