+++ bindgen/src/test/resources/scala-native/struct_passing.h
@@ -3,15 +3,8 @@ typedef struct FunctionArg {
   char *str;
 } FunctionArg;
 
-int function_taking_struct1(FunctionArg arg) { return arg.i * arg.i; }
+int function_taking_struct1(FunctionArg arg);
 
-int function_taking_struct2(FunctionArg arg1, FunctionArg arg2) {
-  return (arg1.i * arg2.i) / 2;
-}
+int function_taking_struct2(FunctionArg arg1, FunctionArg arg2);
 
-FunctionArg function_returning_struct(int value) {
-  // this leaks memory but for tests it should be okay
-  FunctionArg mem = {value * value - 1, "hello"};
-
-  return mem;
-}
+FunctionArg function_returning_struct(int value);
